+++ gradle-suppressible-error-prone/src/test/groovy/com/palantir/gradle/suppressibleerrorprone/SuppressibleErrorPronePluginIntegrationTest.groovy
@@ -315,7 +315,7 @@ class SuppressibleErrorPronePluginIntegrationTest extends IntegrationSpec {
     }
 
     @Unroll
-    def 'blah: #testFile'() {
+    def 'correctly suppresses: #testFile'() {
         def testText = new File('src/test/resources/suppression-test', testFile).text
         def testLines = Splitter.on('\n').splitToList(testText)
 
@@ -353,147 +353,6 @@ class SuppressibleErrorPronePluginIntegrationTest extends IntegrationSpec {
 
         where:
         testFile << Files.list(Path.of("src/test/resources/suppression-test")).map { it.fileName.toString() }.toList()
-    }
-
-    def 'demonstrate suppressions on different source elements'() {
-        // language=Java
-        writeJavaSourceFileToSourceSets '''
-            package app;
-            public final class App {
-                public final String field = new int[3].toString();
-
-                public App() {
-                    System.out.println(new int[3].toString());
-                }
-                
-                public void method() {
-                    System.out.println(new int[3].toString());
-                }
-
-                public void variables() {
-                    String variable = new int[3].toString();
-                    System.out.println(variable);
-                }
-                
-                public static class SomeClass {
-                    static {
-                        System.out.println(new int[3].toString());
-                    }
-                }
-            }
-        '''.stripIndent(true)
-
-        when:
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage1')
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage2')
-
-        then:
-        runTasksSuccessfully('compileAllErrorProne')
-
-        // language=Java
-        appJavaTextEquals '''
-            package app;
-            public final class App {
-                @SuppressWarnings("for-rollout:ArrayToString")
-                public final String field = new int[3].toString();
-                
-                @SuppressWarnings("for-rollout:ArrayToString")
-                public App() {
-                    System.out.println(new int[3].toString());
-                }
-                
-                @SuppressWarnings("for-rollout:ArrayToString")
-                public void method() {
-                    System.out.println(new int[3].toString());
-                }
-
-                public void variables() {
-                    @SuppressWarnings("for-rollout:ArrayToString")
-                    String variable = new int[3].toString();
-                    System.out.println(variable);
-                }
-                
-                @SuppressWarnings("for-rollout:ArrayToString")
-                public static class SomeClass {
-                    static {
-                        System.out.println(new int[3].toString());
-                    }
-                }
-            }
-        '''.stripIndent(true)
-    }
-
-    def 'supports errorprone checks that match on a larger element than they report errors on'() {
-        // The UnusedVariable check implements CompilationUnitTreeMatcher, so will start with a whole
-        // CompilationUnitTree and then narrows down to the specific variable declaration that is unused.
-        // This trips up the "naive" suppression logic, which looks at where the visitor has got to rather
-        // than where the diagnostic description was produced.
-
-        // language=Java
-        writeJavaSourceFileToSourceSets '''
-            package app;
-            public final class App {
-                public void variables() {
-                    String variable;
-                }
-            }
-        '''.stripIndent(true)
-
-        when:
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage1')
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage2')
-
-        then:
-        runTasksSuccessfully('compileAllErrorProne')
-
-        // language=Java
-        appJavaTextEquals '''
-            package app;
-            public final class App {
-                public void variables() {
-                    @SuppressWarnings("for-rollout:UnusedVariable")
-                    String variable;
-                }
-            }
-        '''.stripIndent(true)
-    }
-
-    def 'supports suppressing errorprone checks on classes, interfaces, records, enums, etc'() {
-        // language=Java
-        writeJavaSourceFileToSourceSets '''
-            package app;
-            public final class App {
-                static class exports {}
-                interface opens {}
-                record provides(int cat) {}
-                enum to {;}
-                @interface module {}
-            }
-        '''.stripIndent(true)
-
-        when:
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage1')
-        runTasksSuccessfully('compileAllErrorProne', '-PerrorProneSuppressStage2')
-
-        then:
-        runTasksSuccessfully('compileAllErrorProne')
-
-        // language=Java
-        appJavaTextEquals '''
-            package app;
-            public final class App {
-                @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
-                static class exports {}
-                @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
-                interface opens {}
-                @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
-                record provides(int cat) {}
-                @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
-                enum to {;}
-                @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
-                @interface module {}
-            }
-        '''.stripIndent(true)
     }
 
     def 'can disable errorprone using property'() {
