+++ gradle-suppressible-error-prone/src/test/resources/suppression-test/SuppressesErrorProneChecksThatMatchOnALargerElementThanTheyReportOn.java
@@ -1,0 +1,28 @@
+// Args: -PerrorProneSuppress
+
+// The initial version of this plugin only matched on classes - other "class like" elements did not match.
+// This uses the NamedLikeContextualKeyword to fail on class like elements.
+
+// Before:
+package app;
+public final class App {
+    static class exports {}
+    interface opens {}
+    record provides(int cat) {}
+    enum to {;}
+    @interface module {}
+}
+// After:
+package app;
+public final class App {
+    @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
+    static class exports {}
+    @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
+    interface opens {}
+    @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
+    record provides(int cat) {}
+    @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
+    enum to {;}
+    @SuppressWarnings("for-rollout:NamedLikeContextualKeyword")
+    @interface module {}
+}
