+++ gradle-suppressible-error-prone/src/test/resources/suppression-test/SuppressesFailuresOnClassesInterfacesRecordsEnumsEtc.java
@@ -1,0 +1,16 @@
+// Args: -PerrorProneSuppress
+// Before:
+package app;
+public final class App {
+    public void variables() {
+        String variable;
+    }
+}
+// After:
+package app;
+public final class App {
+    public void variables() {
+        @SuppressWarnings("for-rollout:UnusedVariable")
+        String variable;
+    }
+}
